Dibs on first version!
First attempt to create a keyboard creator based on letter
frequencies. It doesn't do anything smart like looking at letter
combinations or position letters based on anything. It just parses
text and orders it by frequency and then draws that up for a
keyboard layout provided to it.

Example includes material to create an Icelandic keyboard based on
Icelandic parliamentary discussions (what an exciting corpus to
build on).